Move internal components to subfolder
