+++ Source/Extensions/Blazorise.TreeView/Internal/TreeViewNodeState.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Blazorise.TreeView;
+namespace Blazorise.TreeView.Internal;
 
 /// <summary>
 /// Holds the state of the node.
