+++ Source/Extensions/Blazorise.TreeView/TreeView.razor.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Blazorise.Extensions;
+using Blazorise.TreeView.Internal;
 using Microsoft.AspNetCore.Components;
 #endregion
