+++ Source/Extensions/Blazorise.TreeView/_Imports.razor
@@ -1,1 +1,2 @@
 ﻿@using Microsoft.AspNetCore.Components.Web
+@using Blazorise.TreeView.Internal
